Permutation code now works with pre 7.6 MATLAB
Older versions of MATLAB are not compatible with the current way MATLAB
seeds random number generators. When running permutation tests the code
now ignores seeding the random # generator if the version of MATLAB is
too old. This means that test results will not be 100% reproducible,
but if you use a large number of permutations (5000 or more) test
results should still be highly reproducible (just trying running the
test a few times to see how much things change).